update calculation to constants
