+++ src/styles/cropSelectorCalendarView.scss
@@ -13,7 +13,7 @@ div.growthCellsWrapper {
         // margin-top: -7px;
         content: '';
         background-color: $cashCropGrowthWindowColor;
-        // opacity: 100%;
+        opacity: 80%;
         // z-index: -1;
       }
     }
